put the fetching logic in the app.js and add a navigation-bar. just the logic, not the styling.
