+++ components/Layout.js
@@ -1,0 +1,10 @@
+import Navigation from "./Navigation";
+
+export default function Layout({ children }) {
+  return (
+    <>
+      <div>{children}</div>
+      <Navigation />
+    </>
+  );
+}
