+++ pages/art-pieces/index.js
@@ -1,0 +1,5 @@
+import ArtPiecesList from "@/components/ArtPieces";
+
+export default function ArtPiecesPage({ pieces }) {
+  return <ArtPiecesList pieces={pieces} />;
+}
